+++ NEWS.md
@@ -1,4 +1,4 @@
-# glyrepr (development version)
+# glyrepr 0.14.0
 
 ## New features